# api/swagger/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  agent.Agent:
    properties:
      branch_name:
        type: string
      container_id:
        type: string
      created_at:
        type: string
      current_task_id:
        type: string
      health:
        type: string
      id:
        type: string
      process_id:
        type: string
      project_id:
        type: string
      provider:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      worktree_path:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Spec-driven AI development platform — orchestrates teams of Claude
    Code agents.
  title: Foundry API
  version: 0.1.0
paths:
...
  /projects/{id}/agents:
    get:
      description: Returns all agents for the given project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/agent.Agent'
            type: array
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List agents
      tags:
      - agents
  /projects/{id}/agents/{agentId}:
    get:
      description: Returns the agent with the given ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.Agent'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Agent not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get agent
      tags:
      - agents
  /projects/{id}/agents/{agentId}/pause:
    post:
      description: Publishes a pause command for the given agent
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Pause agent
      tags:
      - agents
  /projects/{id}/agents/{agentId}/resume:
    post:
      description: Publishes a resume command for the given agent
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Resume agent
      tags:
      - agents
  /projects/{id}/pause:
    post:
      description: Publishes pause commands for all agents in the project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Pause project
      tags:
      - projects
  /projects/{id}/po/chat:
    delete:
      description: Stub endpoint — PO session manager not yet implemented
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "501":
          description: Not implemented
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: End PO chat (stub)
      tags:
      - po
    post:
      description: Stub endpoint — PO session manager not yet implemented
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "501":
          description: Not implemented
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: PO chat (stub)
      tags:
      - po
  /projects/{id}/po/estimation:
    post:
      description: Stub endpoint — PO session manager not yet implemented
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "501":
          description: Not implemented
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: PO estimation (stub)
      tags:
      - po
  /projects/{id}/po/planning:
    post:
      description: Stub endpoint — PO session manager not yet implemented
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "501":
          description: Not implemented
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: PO planning (stub)
      tags:
      - po
  /projects/{id}/po/status:
    get:
      description: Stub endpoint — returns inactive status until PO session manager
        is implemented
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: PO status (stub)
      tags:
      - po
  /projects/{id}/resume:
    post:
      description: Publishes resume commands for all agents in the project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Resume project
      tags:
      - projects
...
  /projects/{id}/start:
    post:
      description: Initiates project execution (stub — full implementation in Phase
        6)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Start project
      tags:
      - projects
...
produces:
- application/json
schemes:
- http
swagger: "2.0"